backend: Move Regex to config package
